move ai-settings to settings tab
