+++ .prettierignore
@@ -1,1 +1,1 @@
-src/
+# src/
