+++ ai2svelte/src/js/main/Components/TabBar.svelte
@@ -79,9 +79,9 @@
   <div class="container">
     <div class="tab-items">
       {@render tab("Home", true)}
-      {@render tab("CSS", false)}
-      {@render tab("Preview", false)}
       {@render tab("Settings", false)}
+      {@render tab("Styles", false)}
+      {@render tab("Preview", false)}
     </div>
     <div class="theme-configs">
       <ThemeSwitcher bind:theme />
