+++ ai2svelte/src/js/main/Tabs/Styles.svelte
