+++ ai2svelte/src/js/main/state.svelte.ts
@@ -2,7 +2,14 @@ export const userData = $state({
   accentColor: "#dc4300",
   theme: "dark",
   fontsConfig: {
-	Knowledge2017: 'var(--theme-font-family-sans-serif), Knowledge, sans-serif',
-	Knowledge2017TF: 'var(--theme-font-family-sans-serif), Knowledge, sans-serif'
+    Knowledge2017: "var(--theme-font-family-sans-serif), Knowledge, sans-serif",
+    Knowledge2017TF:
+      "var(--theme-font-family-sans-serif), Knowledge, sans-serif",
+    Knowledge2017TF1:
+      "var(--theme-font-family-sans-serif), Knowledge, sans-serif",
+    Knowledge2017TF2:
+      "var(--theme-font-family-sans-serif), Knowledge, sans-serif",
+    Knowledge2017TF3:
+      "var(--theme-font-family-sans-serif), Knowledge, sans-serif",
   },
 });
